chore: rename folders
