+++ App.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { NavigationContainer } from "@react-navigation/native";
 import { API_URL } from "react-native-dotenv";
-import MyStack from "./Navigation/Stacks";
+import MyStack from "./navigation/Stacks";
 
 const App = () => {
   return (
