+++ components/LoadUrl.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import { WebView } from "react-native-webview";
+import { View, ActivityIndicator } from "react-native";
 
 type Props = {
   handleDetails: (event: any) => Promise<void>;
@@ -9,11 +10,8 @@ type Props = {
 const LoadUrl = ({ handleDetails, etUrl }: Props) => {
   const injectJavascript = `
   document.addEventListener('click', function(event) {
-    const clickedOnTree = document.getElementsByClassName('greenstand-point-selected-box green-b');
-    if(clickedOnTree){
       const data =  Object.values(document.querySelector("#map-canvas").map?.map._layers).reduce((a,c) => a || c.payload?.active?c.payload:undefined, undefined)
         window.ReactNativeWebView.postMessage(JSON.stringify(data));
-    }
    })
   `;
 
@@ -22,6 +20,20 @@ const LoadUrl = ({ handleDetails, etUrl }: Props) => {
       source={{ uri: etUrl }}
       injectedJavaScript={injectJavascript}
       onMessage={handleDetails}
+      startInLoadingState={true}
+      renderLoading={() => (
+        <ActivityIndicator
+          style={{
+            position: "absolute",
+            width: "100%",
+            height: "100%",
+            justifyContent: "center",
+            alignItems: "center",
+          }}
+          size="large"
+          color="green"
+        />
+      )}
     />
   );
 };
